+++ server/interaction-api/src/main/java/dns/vitrina/dto/UserAuthDto.java
@@ -1,0 +1,15 @@
+package dns.vitrina.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class UserAuthDto {
+    @NotBlank
+    @Size(min = 5)
+    private String lastName;
+    @NotBlank
+    @Size(min = 5)
+    private String password;
+}
