+++ server/interaction-api/src/main/java/dns/vitrina/dto/UserDto.java
@@ -1,0 +1,15 @@
+package dns.vitrina.dto;
+
+import dns.vitrina.server.model.Task;
+import lombok.Data;
+
+import java.util.List;
+
+
+@Data
+public class UserDto {
+    Long id;
+    String firstName;
+    String lastName;
+    List<Task> tasks;
+}
