+++ server/interaction-api/src/main/java/dns/vitrina/dto/VitrinaDto.java
@@ -1,0 +1,16 @@
+package dns.vitrina.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class VitrinaDto {
+    private Long id;
+    private String name;
+    private int x;
+    private int y;
+    private int width;
+    private int height;
+    private List<UserDto> usersToVitrina;
+}
