+++ server/interaction-api/src/main/java/dns/vitrina/exception/AddedUserException.java
@@ -1,0 +1,5 @@
+package dns.vitrina.exception;
+
+public class AddedUserException extends RuntimeException{
+    public AddedUserException(String message) {super(message);}
+}
