+++ server/interaction-api/src/main/java/dns/vitrina/exception/NotDateBaseUserException.java
@@ -1,0 +1,5 @@
+package dns.vitrina.exception;
+
+public class NotDateBaseUserException extends RuntimeException{
+    public NotDateBaseUserException(String message) {super(message);}
+}
