--- server/src/main/java/dns/vitrina/server/dto/UserAuthDto.java
@@ -1,15 +1,0 @@
-package dns.vitrina.server.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-
-@Data
-public class UserAuthDto {
-    @NotBlank
-    @Size(min = 5)
-    private String lastName;
-    @NotBlank
-    @Size(min = 5)
-    private String password;
-}
