--- server/src/main/java/dns/vitrina/server/dto/UserDto.java
@@ -1,15 +1,0 @@
-package dns.vitrina.server.dto;
-
-import dns.vitrina.server.model.Task;
-import lombok.Data;
-
-import java.util.List;
-
-
-@Data
-public class UserDto {
-    Long id;
-    String firstName;
-    String lastName;
-    List<Task> tasks;
-}
