--- server/src/main/java/dns/vitrina/server/exception/AddedUserException.java
@@ -1,5 +1,0 @@
-package dns.vitrina.server.exception;
-
-public class AddedUserException extends RuntimeException{
-    public AddedUserException(String message) {super(message);}
-}
