--- server/src/main/java/dns/vitrina/server/exception/NotDateBaseUserException.java
@@ -1,5 +1,0 @@
-package dns.vitrina.server.exception;
-
-public class NotDateBaseUserException extends RuntimeException{
-    public NotDateBaseUserException(String message) {super(message);}
-}
