--- server/src/main/java/dns/vitrina/server/model/StatusTask.java
@@ -1,7 +1,0 @@
-package dns.vitrina.server.model;
-
-public enum StatusTask {
-    PENDING,
-    SUCCESS,
-    NOT_DONE
-}
