--- server/src/main/java/dns/vitrina/server/model/Task.java
@@ -1,37 +1,0 @@
-package dns.vitrina.server.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.time.LocalDate;
-import java.util.Date;
-
-@Data
-@Entity
-@Inheritance(strategy = InheritanceType.JOINED)
-@Table(name = "tasks")
-public class Task {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column
-    private Long id;
-
-    @Column
-    private String name;
-
-    @Column
-    private String description;
-
-    @Column
-    private StatusTask status;
-
-    @Column
-    private LocalDate created;
-
-    @Column
-    private LocalDate endDate;
-
-    @Column
-    private Long vitrinaId;
-}
