--- server/src/main/java/dns/vitrina/server/model/User.java
@@ -1,29 +1,0 @@
-package dns.vitrina.server.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.List;
-
-@Entity
-@Data
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column
-    private Long id;
-
-    @Column
-    private String firstName;
-
-    @Column
-    private String lastName;
-
-    @Column
-    private String password;
-
-    @OneToMany
-    private List<Task> tasks;
-}
