--- server/src/main/java/dns/vitrina/server/repository/TaskRepository.java
@@ -1,10 +1,0 @@
-package dns.vitrina.server.repository;
-
-import dns.vitrina.server.model.Task;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
-
-}
