--- server/src/main/java/dns/vitrina/server/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package dns.vitrina.server.repository;
-
-import dns.vitrina.server.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByLastName(String lastName);
-    boolean findByLastName(String lastName);
-}
