--- server/src/main/java/dns/vitrina/server/repository/VitrinaRepository.java
@@ -1,9 +1,0 @@
-package dns.vitrina.server.repository;
-
-import dns.vitrina.server.model.Vitrina;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface VitrinaRepository extends JpaRepository<Vitrina, Long> {
-}
