+++ server/task-service/src/main/java/dns/vitrina/TaskApplication.java
@@ -1,0 +1,13 @@
+package dns.vitrina;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class TaskApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TaskApplication.class, args);
+    }
+}
