+++ server/task-service/src/main/java/dns/vitrina/model/StatusTask.java
@@ -1,0 +1,7 @@
+package dns.vitrina.model;
+
+public enum StatusTask {
+    PENDING,
+    SUCCESS,
+    NOT_DONE
+}
