+++ server/task-service/src/main/java/dns/vitrina/repository/TaskRepository.java
@@ -1,0 +1,10 @@
+package dns.vitrina.repository;
+
+import dns.vitrina.server.model.Task;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TaskRepository extends JpaRepository<Task, Long> {
+
+}
