+++ server/user-service/src/main/java/dns/vitrina/UserApplication.java
@@ -1,0 +1,11 @@
+package dns.vitrina;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class UserApplication {
+    public static void main(String[] args) {SpringApplication.run(UserApplication.class, args);}
+}
