+++ server/user-service/src/main/java/dns/vitrina/mapper/UserMapper.java
@@ -1,0 +1,26 @@
+package dns.vitrina.mapper;
+
+
+import dns.vitrina.dto.UserAuthDto;
+import dns.vitrina.dto.UserDto;
+import dns.vitrina.dto.UserRequest;
+import dns.vitrina.model.User;
+import org.mapstruct.Mapper;
+import org.mapstruct.MappingConstants;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+public interface UserMapper {
+    UserDto userToUserDto(User user);
+
+    List<UserDto> usersToUserDtos(List<User> users);
+
+    User userDtoToUserCreate(UserRequest userDto);
+
+    UserAuthDto userToUserAuthDto(User user);
+
+    User userAuthDtoToUser(UserAuthDto userAuthDto);
+}
