+++ server/vitrina-service/src/main/java/dns/vitrina/service/VitrinaService.java
@@ -1,0 +1,15 @@
+package dns.vitrina.service;
+
+import dns.vitrina.server.dto.VitrinaDto;
+import dns.vitrina.server.model.Vitrina;
+
+import java.util.List;
+
+public interface VitrinaService {
+
+    Vitrina get(Long id);
+
+    List<VitrinaDto> getAll();
+
+    void update(Vitrina vitrina);
+}
